refactor: notification_receiverパッケージ内のProviderをGenerator Providerに置き換え
