+++ packages/notification_receiver/lib/notification_receiver.dart
@@ -4,6 +4,5 @@ library notification_receiver;
 export 'package:firebase_messaging/firebase_messaging.dart'
     show NotificationSettings;
 
-export 'src/notification_configurator.dart';
+export 'src/notification_configurator_provider.dart';
 export 'src/notification_settings_extension.dart';
-export 'src/notification_settings_provider.dart';
