+++ packages/notification_receiver/lib/src/firebase_messaging_provider.dart
@@ -1,4 +1,8 @@
 import 'package:firebase_messaging/firebase_messaging.dart';
-import 'package:riverpod/riverpod.dart';
+import 'package:riverpod_annotation/riverpod_annotation.dart';
 
-final firebaseMessagingProvider = Provider((_) => FirebaseMessaging.instance);
+part 'firebase_messaging_provider.g.dart';
+
+@Riverpod(keepAlive: true)
+FirebaseMessaging firebaseMessaging(FirebaseMessagingRef ref) =>
+    FirebaseMessaging.instance;
